+++ src/main/java/com/schnabel/schnabel/pharmacies/service/IPharmacyService.java
@@ -3,6 +3,6 @@ package com.schnabel.schnabel.pharmacies.service;
 import com.schnabel.schnabel.misc.interfaces.ICrudService;
 import com.schnabel.schnabel.pharmacies.model.Pharmacy;
 
-public interface IPharmacyService extends ICrudService<Pharmacy, Integer>
+public interface IPharmacyService extends ICrudService<Pharmacy, Long>
 {
 }
